+++ packages/chain-adapters/src/bitcoin/BitcoinChainAdapter.ts
@@ -127,7 +127,7 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
   }
 
   async buildSendTransaction(
-    tx: chainAdapters.BuildSendTxInput
+    tx: chainAdapters.BuildSendTxInput<ChainTypes.Bitcoin>
   ): Promise<{
     txToSign: chainAdapters.ChainTxType<ChainTypes.Bitcoin>
   }> {
@@ -137,8 +137,7 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
         to,
         wallet,
         bip32Params = ChainAdapter.defaultBIP32Params,
-        feeSpeed,
-        scriptType = BTCInputScriptType.SpendWitness
+        chainSpecific: { satoshiPerByte, scriptType }
       } = tx
 
       if (!value || !to) {
@@ -169,8 +168,6 @@ export class ChainAdapter implements IChainAdapter<ChainTypes.Bitcoin> {
       if (!from) throw new Error('BitcoinChainAdapter: from undefined')
 
       const account = await this.getAccount(pubkey.xpub)
-      const estimatedFees = await this.getFeeData({ to, value, from })
-      const satoshiPerByte = estimatedFees[feeSpeed ?? chainAdapters.FeeDataKey.Average].feePerUnit
 
       type MappedUtxos = Omit<bitcoin.api.Utxo, 'value'> & { value: number }
       const mappedUtxos: MappedUtxos[] = utxos.map((x) => ({ ...x, value: Number(x.value) }))